pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-factory"
version = "0.1.2"
description = "The LEGO set for custom vLLM model plugins — build, test, and deploy custom encoders, poolers, and kernels"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "LatenceAI", email = "hello@latence.ai"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "sentencepiece",
    "huggingface-hub",
    "pillow",
    "requests",
]

[project.optional-dependencies]
gliner = [
    "gliner>=0.2.26",
]
glinker = [
    "gliner>=0.2.26",
    "glinker>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-benchmark",
    "ruff>=0.4",
    "numpy",
    "packaging>=24.0",
]
kernels = [
    "triton>=2.0",
]

[project.scripts]
vllm-factory-prep = "forge.__main__:main"

[project.entry-points."vllm.general_plugins"]
embeddinggemma = "plugins.embeddinggemma:register"
mt5_gliner = "plugins.mt5_gliner:register"
mmbert_gliner = "plugins.mmbert_gliner:register"
deberta_gliner = "plugins.deberta_gliner:register"
deberta_gliner2 = "plugins.deberta_gliner2:register"
deberta_gliner_linker = "plugins.deberta_gliner_linker:register"
modernbert_gliner_rerank = "plugins.modernbert_gliner_rerank:register"
moderncolbert = "plugins.moderncolbert:register"
colqwen3 = "plugins.colqwen3:register"
collfm2 = "plugins.collfm2:register"
nemotron_colembed = "plugins.nemotron_colembed:register"
lfm2_colbert = "plugins.lfm2_colbert:register"

[project.entry-points."vllm.io_processor_plugins"]
mmbert_gliner_io = "plugins.mmbert_gliner.io_processor:get_processor_cls"
colqwen3_io = "plugins.colqwen3.io_processor:get_processor_cls"
deberta_gliner_io = "plugins.deberta_gliner.io_processor:get_processor_cls"
mt5_gliner_io = "plugins.mt5_gliner.io_processor:get_processor_cls"
deberta_gliner2_io = "plugins.deberta_gliner2.io_processor:get_processor_cls"
moderncolbert_io = "plugins.moderncolbert.io_processor:get_processor_cls"
collfm2_io = "plugins.collfm2.io_processor:get_processor_cls"
deberta_gliner_linker_io = "plugins.deberta_gliner_linker.io_processor:get_processor_cls"
modernbert_gliner_rerank_io = "plugins.modernbert_gliner_rerank.io_processor:get_processor_cls"
nemotron_colembed_io = "plugins.nemotron_colembed.io_processor:get_processor_cls"
lfm2_colbert_io = "plugins.lfm2_colbert.io_processor:get_processor_cls"
embeddinggemma_io = "plugins.embeddinggemma.io_processor:get_processor_cls"

[project.urls]
Homepage = "https://github.com/ddickmann/vllm-factory"
Documentation = "https://github.com/ddickmann/vllm-factory/tree/main/docs"
Repository = "https://github.com/ddickmann/vllm-factory"
Issues = "https://github.com/ddickmann/vllm-factory/issues"

[tool.setuptools.packages.find]
include = ["forge*", "models*", "poolers*", "kernels*", "processors*", "plugins*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: marks tests as requiring GPU (deselect with '-m \"not gpu\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
